include/nutconf.hpp: align UpsmonConfiguration::notifyFlags type with clients/upsmon.h (notifylist->flags)
Note that here we have "unsigned int" and there a plain "int" at the moment;
but even with short signed ints we currently do not have too many values :)

This fix also drops a warning about shortening the cast when assigning from
a temporary "unsigned int flags" in UpsmonConfigParser::onParseDirective()

+++ include/nutconf.hpp
@@ -664,7 +664,7 @@ public:
 	static NotifyFlag NotifyFlagFromString(const std::string& str);
 	static NotifyType NotifyTypeFromString(const std::string& str);
 
-	Settable<unsigned short> notifyFlags[NOTIFY_TYPE_MAX];
+	Settable<unsigned int>   notifyFlags[NOTIFY_TYPE_MAX];
 	Settable<std::string>    notifyMessages[NOTIFY_TYPE_MAX];
 
 	struct Monitor {
